Struct screenprints::Printer

pub struct Printer {
    // some fields omitted
}

A Printer buffers writes and flushes at a specified interval, clearing the display of any lines of text previously written

Methods

impl Printer

fn new<W>(underlying: W, interval: Duration) -> Printer where W: Write + Send + 'static

Creates a new Printer instance that delegates writes to the provided Write instance delayed at the interval provided

fn clear(&self)

clear the current buffer and reset linecount

fn close(&self)

close the printer, afterwhich all writes will be discarded